exam/requirements.txt
flask
pytesseract
pillow
pdf2image
sentence-transformers
gpt4all
fpdf
